Move FilterComponent to top level
This means we can use it in other namespaces.
