+++ app/views/provider_interface/application_choices/index.html.erb
@@ -21,7 +21,7 @@
 <h1 class='govuk-heading-xl'>Applications</h1>
 
 <div class='moj-filter-layout'>
-  <%= render ProviderInterface::FilterComponent.new(page_state: @page_state) %>
+  <%= render FilterComponent.new(page_state: @page_state) %>
   <div class='moj-filter-layout__content'>
     <% if @application_choices.any? %>
       <div class="app-application-cards">
